refactor: move identifier.rs tests to separate file
Move the 2172-line test module from identifier.rs into
identifier/tests.rs, leaving 765 lines of implementation
in identifier/mod.rs.

No logic changes - pure structural refactoring. All 51
identifier tests pass unchanged.

Refs #1850
